// WCI/src/app/models/equipmentType.ts
export enum EquipmentType
{
    Unknown = 0,
    Instrument = 1,                // managed instruments
    Furniture = 2,                // furnitures used in Bridge view
    Storage = 3,                // storage/Refrigerators
    UnmanagedInstrument = 4,    // instruments that are currently not managed by Bridge
    LISTerminal = 5,            // LIS terminal
    NetworkedDevice = 6,
    AutomationModule = 7,
    LPMServer = null
};